Start implementation of support for Link Images
Started implementation / support for link images, first getting the OpenGraph Image if it is set, then falling back to the first image on the page, if one is available.

For Youtube, they provide a unique URL for max resolution Youtube thumbnails, so we'll utilize that alongside the video ID instead.