Improve YouTube parsing.
For YouTube, default to disabling polymer so playlist URLs behave better, before passing it to our YouTube parser. Once we do this, if we fail to get a title for the page (like a video) then attempt to fetch the meta itemprop name content value for the title or fallback to just having the title be "YouTube".